remaining needs (and should only) be a long. (ap_get_chunk_size) PR: Obtained from: Submitted by: Reviewed by:
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@95728 13f79535-47bb-0310-9956-ffa450edef68

+++ src/modules/proxy/proxy_util.c
@@ -452,7 +452,7 @@ long int ap_proxy_send_fb(BUFF *f, request_rec *r, cache_req *c, off_t len, int 
     int ok, end_of_chunk;
     char *buf;
     size_t buf_size;
-    size_t remaining = 0;
+    long remaining = 0;
     long total_bytes_rcvd;
     register int n = 0, o, w;
     conn_rec *con = r->connection;
